+++ packages/cli/src/generator/test-report/index.ts
@@ -13,10 +13,10 @@ export default class Report extends Generator {
 
 	writing() {
 		this.fs.copyTpl(this.templatePath('report.html'), this.destinationPath('report.html'), {})
-		this.fs.copyTpl(this.templatePath('styles.css'), this.destinationPath('styles.css'), {})
-		this.fs.copyTpl(this.templatePath('js'), this.destinationPath('js'), {})
-		this.fs.write(this.destinationPath('js/data.js'), `var data = ${this.options.data}`)
 		this.fs.copyTpl(this.templatePath('assets'), this.destinationPath('assets'), {})
+		this.fs.move(this.templatePath('styles.css'), this.destinationPath('styles.css'), {})
+		this.fs.move(this.templatePath('script.js'), this.destinationPath('js/script.js'), {})
+		this.fs.write(this.destinationPath('js/data.js'), `var data = ${this.options.data}`)
 		this.fs.write(this.destinationPath('data.json'), this.options.data)
 	}
 }