+++ packages/core/src/page/types.ts
@@ -278,3 +278,11 @@ export interface ScreenshotOptions {
 	 */
 	encoding?: 'base64' | 'binary'
 }
+
+export interface CookiesFilterParams {
+	urls?: string | string[]
+	names?: string | string[]
+}
+
+// Using the type below for API scrollTo()
+export type ScrollDirection = 'top' | 'left' | 'bottom' | 'right'