+++ CHANGELOG.md
@@ -3,6 +3,94 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-canary.20](https://github.com/flood-io/element/compare/v2.0.0-canary.19...v2.0.0-canary.20) (2021-01-12)
+
+
+### Bug Fixes
+
+* fix codacy ([95823e7](https://github.com/flood-io/element/commit/95823e79223f17b9c36320255c2b8039dc49fe39))
+* refactor some lines of code ([ca2b7b3](https://github.com/flood-io/element/commit/ca2b7b37d6e98c0701f54be4f9d3d515e2d3ac76))
+* remove redundant code ([7d8b07e](https://github.com/flood-io/element/commit/7d8b07e24a582c17d51f4d1a139e2ea14567792e))
+* update message of test script template ([ab31c32](https://github.com/flood-io/element/commit/ab31c32cb93e6f95452574d571f8a82380b3e60f))
+
+
+### Features
+
+* implement API getUrl() ([bc6ad6e](https://github.com/flood-io/element/commit/bc6ad6ef7d553be4e42f4e4bb7f49f237971a60a))
+* implement browser.scrollBy() from 1.4 to 2.0 ([bc0c57b](https://github.com/flood-io/element/commit/bc0c57b6223d19f0b5a656c1496ffa599ead474e))
+* implement browser.scrollTo(), browser.scrollBy() from 1.4 to 2.0 ([89dfb9c](https://github.com/flood-io/element/commit/89dfb9c9081d5d7adc6f43a34dd217291872b3c2))
+* print version of Element and Node when running ([b3d63eb](https://github.com/flood-io/element/commit/b3d63eb020043dba253e833380c4ac550fc130f6))
+
+
+
+
+
+# [2.0.0-canary.19](https://github.com/flood-io/element/compare/v2.0.0-canary.18...v2.0.0-canary.19) (2021-01-11)
+
+
+### Bug Fixes
+
+* remove highlightElement api because playwright does not support yet ([3aeee26](https://github.com/flood-io/element/commit/3aeee26f0c9cdbe650611ece644d762ec3945377))
+
+
+### Features
+
+* add hooks to template ([ff17558](https://github.com/flood-io/element/commit/ff175585203103fc734c30b6f62faaa47e9821a5))
+
+
+
+
+
+# [2.0.0-canary.18](https://github.com/flood-io/element/compare/v2.0.0-canary.17...v2.0.0-canary.18) (2021-01-08)
+
+
+### Bug Fixes
+
+* cannot find module '../../package.json' ([42ed4e6](https://github.com/flood-io/element/commit/42ed4e6d32bca65904e11a817361ef5a81265cd7))
+
+
+
+
+
+# [2.0.0-canary.17](https://github.com/flood-io/element/compare/v2.0.0-canary.16...v2.0.0-canary.17) (2021-01-07)
+
+
+### Bug Fixes
+
+* errors reported by codacy ([1a41918](https://github.com/flood-io/element/commit/1a41918b5cf89cd3e9ac2a81db26ac70d533589a))
+* errors reported by codacy ([3246406](https://github.com/flood-io/element/commit/32464067b05fa39378570a63bdbfb6827049204d))
+* errors reported by codacy ([6586df6](https://github.com/flood-io/element/commit/6586df62d008164950ae5d9e05aa060ca4741e37))
+* errors reported by codacy ([bb52432](https://github.com/flood-io/element/commit/bb524322c3c7ff2d2f84c312f7dabd3d61612f3c))
+* errors reported by codacy ([4cc2269](https://github.com/flood-io/element/commit/4cc22693975672c555d2ff1d960dbd78b4f273ab))
+* report table ([3e042dd](https://github.com/flood-io/element/commit/3e042ddc858f6921ef6ce112ea9e49da3da3e5fb))
+* update unit test ([e3be066](https://github.com/flood-io/element/commit/e3be06600910c40f506ed181dfecf8d2b8cff3e4))
+
+
+### Features
+
+* allow filtering steps by result ([29fd092](https://github.com/flood-io/element/commit/29fd092a8897a6f46d89f67ac4bd46ef12ceeec9))
+* implement ElementCookies ([f728792](https://github.com/flood-io/element/commit/f728792ae1eb7242daa6699348fafb77e57df431))
+* refactor code ([06b1950](https://github.com/flood-io/element/commit/06b19502ddf643f390057c493a3fbf28ae507e81))
+* refactor code base on suggestion of reviewers ([60c77ac](https://github.com/flood-io/element/commit/60c77ac1a225165539d22c85be36976c3059963a))
+* update API base on new AC ([81c1c53](https://github.com/flood-io/element/commit/81c1c536080975f43798180f5601d2127758ef24))
+* update unit test ([1e0d52b](https://github.com/flood-io/element/commit/1e0d52bbaf1be4a19db0644735dc30e355f93677))
+* view summary in number or percentage ([7d00010](https://github.com/flood-io/element/commit/7d0001070f834a2e1709cf1c34145d62632cfdea))
+
+
+
+
+
+# [2.0.0-canary.16](https://github.com/flood-io/element/compare/v2.0.0-canary.15...v2.0.0-canary.16) (2021-01-07)
+
+
+### Bug Fixes
+
+* refactor the way check file arg exist or not ([65e064e](https://github.com/flood-io/element/commit/65e064e76ceec0532a7b9fe3eee01546260910ca))
+
+
+
+
+
 # [2.0.0-canary.15](https://github.com/flood-io/element/compare/v2.0.0-canary.14...v2.0.0-canary.15) (2021-01-06)